obtain_clone_cpuid(): always start from CPU 0 and fill in cores linearily

+++ kernel/process.c
@@ -2325,7 +2325,7 @@ redo:
 		struct cpu_local_var *v;
 		struct cpu_local_var *cur_v;
 		struct process *proc_to_move = NULL;
-		int irqstate2;
+		unsigned long irqstate2;
 
 		irqstate = cpu_disable_interrupt_save();
